requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.1
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djangorestframework-simplejwt-sso==1.0.1
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
